+++ src/Geometry/MeshFactory.ts
@@ -1,4 +1,3 @@
-import { Scalar } from './../Math';
 import { MeshProcessor } from './MeshProcessor';
 import { IMesh } from "./Mesh";
 
@@ -10,14 +9,9 @@ export enum PlatonicSolids {
     icosahedron
 }
 
-export type SubdivisionPattern = 4 | 9;
-
-
 export class MeshFactory {
 
     public static DefaultRadius:number = 1.0;
-    public static DefaultSubdivisionLevel:number = 4;
-    public static DefaultLoopPattern:SubdivisionPattern = 4;
     public static DefaultSubdivisionLevelA: number = 8;
  
     /* The following mesh is based on the treatment in Geometric Tools for Computer Graphics (Morgan Kaufmann 2003) chapter 9.3.6 p 346-350
@@ -173,10 +167,10 @@ export class MeshFactory {
             let p1 = shape.indices[fi++];
             let p2 = shape.indices[fi++];
 
-            let vedge0 = MeshFactory._divideEdge(shape,p1,p0,a,radius,map);
-            let vedge1 = MeshFactory._divideEdge(shape,p1,p2,a,radius,map);
+            let vedge0 = MeshProcessor.DivideEdge(shape,p1,p0,a,radius,map);
+            let vedge1 = MeshProcessor.DivideEdge(shape,p1,p2,a,radius,map);
 
-            let hedges = vedge0.map((value,index,array)=>MeshFactory._divideEdge0(shape,value,vedge1[index],index,radius));
+            let hedges = vedge0.map((value,index,array)=>MeshProcessor.DivideEdge(shape,value,vedge1[index],index,radius,undefined));
             
             let row0: number[],row1: number[];
             
@@ -184,247 +178,17 @@ export class MeshFactory {
                 row0 = hedges[i++];
                 row1 = hedges[i  ];
 
-                MeshFactory._pushFace(row1[0],row0[0],row1[1],newFaces,shape.uvs[0]);
+                newFaces.push(row1[0],row0[0],row1[1]);
 
                 for(let k=1;k<row1.length-1;k++){
-                    MeshFactory._pushFace(row1[k],row0[k-1],row0[k  ],newFaces,shape.uvs[0]);
-                    MeshFactory._pushFace(row1[k],row0[k  ],row1[k+1],newFaces,shape.uvs[0]);
+                    newFaces.push(row1[k],row0[k-1],row0[k  ]);
+                    newFaces.push(row1[k],row0[k  ],row1[k+1]);
                 }
             }
         }
-        
+
         shape.indices = newFaces;
 
         return shape;
-    }
-
-    private static _pushFace(a:number,b:number,c:number,faces: number[], uvs: number[]){
-        faces.push(a,b,c);
-    }
-
-
-    private static _divideEdge(shape:IMesh, p0:number, p1:number, a:number, radius:number, map :{[key in string]:number[]}) : number[] {
- 
-        var key:string = p0 + '_' + p1;
-        var list = map[key];
-        if( list ) {
-            return list;
-        }
-        var inversedkey = p1 + '_' + p0;
-        list = map[inversedkey];
-        if( list ) {
-            return list.slice().reverse();
-        }
-        list = MeshFactory._divideEdge0(shape,p0,p1,a,radius);
-        map[key] = list;
-        return list;
-    }
-
-    private static _divideEdge0(shape:IMesh, p0:number, p1:number, a:number, radius:number) : number[] {
- 
-        let list = [p0];
-        if(a != 0  && p0 != p1 ){
-            if(a > 1){
-                let vertices:number[] = shape.vertices;
-                let normals:number[] = shape.normals;
-                let uvs:number[] = shape.uvs[0];
-                let i0x = p0*3;
-                let i0y = i0x+1;
-                let i0z = i0x+2;
-
-                let i1x = p1*3;
-                let i1y = i1x+1;
-                let i1z = i1x+2;
-
-                let dx = (vertices[i1x]-vertices[i0x])/a;
-                let dy = (vertices[i1y]-vertices[i0y])/a;
-                let dz = (vertices[i1z]-vertices[i0z])/a;
-
-                let n = vertices.length/3 ;
-
-                for(let i=1; i!= a; i++) {
-                    let x:number = vertices[i0x] + dx*i;
-                    let y:number = vertices[i0y] + dy*i;
-                    let z:number = vertices[i0z] + dz*i;
-
-                    let l:number = Math.sqrt(x*x+y*y+z*z);
-                    x/=l;
-                    y/=l;
-                    z/=l;
-
-                    list.push(n++);
-                    vertices.push(x*radius,y*radius,z*radius);
-                    normals.push(x,y,z);
-
-                    /*  
-                        vertices are not distributed evenly across longitude and latitude then we MUST compute the uv with formula
-                        theta is from +/-[0-PI] 
-                        phi is from  [0-PI]
-                    */  
-                    let theta = Math.atan2(z,x);
-                    let phi = Math.acos(y) ; 
-                    let v:number = 1 - phi/Math.PI ; // normalize 
-                    let u:number = theta / (2*Math.PI) ; // normalize
-                    u = theta >= 0 ?  u : 1 + u ;
-                    u = Scalar.WithinEpsilon(u,0)?0:u;
-
-                    // avoid zip effect : 
-                    let u1 = uvs[p0*2];
-                    let u2 = uvs[p1*2]
- 
-                    let d1 = u1 - u;
-                    let d2 = u2 - u;
-                    if( Math.abs(d1) > .5 || Math.abs(d2) > .5 ) {
-                        u+=1;
-                    }
-                    
-                    uvs.push(u,v);
-                }
-            }
-            list.push(p1);
-        }
-        return list;
-    }
-    /* 
-        icosphere is a particular simple tesselation method where we project subdivided points to the sphere surface. 
-        then we DO NOT relay on memory/compute intensive subdivision method.
-        TODO : because of incremental error propagation, we might consider to find analytic approach for vertex coordinate generation instead as parametric subdivision methods.
-    */
-    public static CreateIcosphere(shape: IMesh, radius:number, subdivisionLevel : number =  MeshFactory.DefaultSubdivisionLevel, pattern:SubdivisionPattern = MeshFactory.DefaultLoopPattern ): IMesh {
-
-       radius = Math.abs(radius || MeshFactory.DefaultRadius);
-       shape = shape = MeshFactory._createIcosphereBase(shape || <IMesh>{}, radius);
-
-       subdivisionLevel = subdivisionLevel === undefined ? MeshFactory.DefaultSubdivisionLevel : Math.abs(subdivisionLevel);
-
-       pattern = pattern == undefined ? MeshFactory.DefaultLoopPattern : pattern != 4 && pattern != 9 ? MeshFactory.DefaultLoopPattern : pattern; /* ensure patternn is 4 OR 9 */
-       for(let i=0;i!= subdivisionLevel;i++) {
-            
-            let map :{[key in string]:number} = {}
-            let newFaces: number[] = []
-            let n = shape.indices.length;
-            
-            for(let f=0; f !== n;) {
-
-                let p0 = shape.indices[f++];
-                let p1 = shape.indices[f++];
-                let p2 = shape.indices[f++];
-
-                switch(pattern){
-                    case(9):
-                    {
-                               /* 0
-                                 / \
-                                /   \
-                               a --- g
-                              / \   / \
-                             /   \ /   \
-                            d --- c --- f     
-                           / \   / \   / \
-                          /   \ /   \ /   \
-                         1 --- e --- b --- 2 */
-                        let t13:number = 1.0/3;
-                        let t23:number = 2.0/3;
-                        let t12:number = .5;
-                        let a = MeshFactory._getParametricPoint(shape,p0,p1,t13,radius,map);
-                        let b = MeshFactory._getParametricPoint(shape,p1,p2,t23,radius,map);
-                        let c = MeshFactory._getParametricPoint(shape,a ,b ,t12,radius,map);
-                        let d = MeshFactory._getParametricPoint(shape,p0,p1,t23,radius,map);
-                        let e = MeshFactory._getParametricPoint(shape,p1,p2,t13,radius,map); 
-                        let f = MeshFactory._getParametricPoint(shape,p2,p0,t13,radius,map);
-                        let g = MeshFactory._getParametricPoint(shape,p2,p0,t23,radius,map);
-                        newFaces.push(p0, a, g, a, d, c, a, c, g, g, c, f, d, p1, e, d, e, c, c, e, b, c, b, f, f, b, p2);
-                        break;
-                    }
-                    case(4):
-                    default : {
-                               /* 0
-                                 / \
-                                /   \
-                               a --- c
-                              / \   / \
-                             /   \ /   \
-                            1 --- b --- 2 */    
-                        let t:number = 0.5;
-                        let a = MeshFactory._getParametricPoint(shape,p0,p1,t,radius,map);
-                        let b = MeshFactory._getParametricPoint(shape,p1,p2,t,radius,map);
-                        let c = MeshFactory._getParametricPoint(shape,p2,p0,t,radius,map);
-                        newFaces.push(p0, a, c, a, p1, b, a, b, c, c, b, p2);
-                    }
-                }
-            }
-            shape.indices = newFaces;
-        }
-        return shape;
-    }
-
-    private static _getParametricPoint(shape: IMesh, p1:number, p2:number, t:number, radius:number, map :{[key in string]:number}) : number {
-        
-        /* IS the tricks to share the same point when divide top or bottom segment */
-        var k1:string = p1 < 10 ? "*" : p1.toString(); 
-        var k2:string = p2 < 10 ? "*" : p2.toString(); 
-
-        var key:string = (p1<p2? k1 + "-"+  k2: k2 + "-"+  k1) + "_" + t ;
-        var i = map[key];
-        if( i ) {
-            delete map[key]; /* avoid too many key */
-            return i;
-        }
-        let i1 = p1*3;
-        let i2 = p2*3;
-        let v:number[] = shape.vertices;
-        let x:number = v[i1] + (v[i2++] - v[i1++]) * t ;
-        let y:number = v[i1] + (v[i2++] - v[i1++]) * t ; 
-        let z:number = v[i1] + (v[i2  ] - v[i1  ]) * t ;
-
-        let l:number = Math.sqrt(x*x+y*y+z*z);
-        x/=l;
-        y/=l;
-        z/=l;
-
-        let n:number[] = shape.normals;
-        n.push(x,y,z);
- 
-        i = v.length/3;
-        map[key] = i;
-
-        v.push(x*radius,y*radius,z*radius);
-
-        i1 = p1*2;
-        i2 = p2*2;
-
-        /* 
-           vertices are not distributed evenly across longitude and latitude then we MUST compute the uv with formula
-           theta is from +/-[0-PI] 
-           phi is from  [0-PI]
-        */  
-        let theta = Math.atan2(z,x);
-        let phi = Math.acos(y) ; 
-
-        for(let j=0;j!= shape.uvs.length; j++){
-            let uvs:number[] = shape.uvs[j];
-
-            let u1 = uvs[i1++];
-            let u2 = uvs[i2++]
-            let v1 = uvs[i1];
-            let v2 = uvs[i2]; 
-
-            let a = theta / (2*Math.PI) ;
-            a = Scalar.WithinEpsilon(a,0)?0:a;
-
-            let u:number = theta >= 0 ?  a : 1 + a ;
-            let d1 = u1 - u;
-            let d2 = u2 - u;
-            if( Math.abs(d1) > .5 || Math.abs(d2) > .5 ) {
-                u+=1;
-            }
-           
-
-            let v:number = 1 - phi/Math.PI ;
-
-            uvs.push(u,v);
-        }
- 
-        return i;
     }
 }
